Another Big one
the table is fully implemented, I am now working on debugging issues
with the parse function, based on the fact that certain words such as
'r343r' are not valid as whole words but are valid as words of a
variable I had to do a complete work around of the function based if it
would read a full word or character by character. Still bug testing,
will have more done hopefully tomorrow.